+++ docs/docs/CHANGELOG.md
@@ -2,13 +2,13 @@
 
 ## v0.9.1
 
-## Changes
+### Changes
 - Remove unnecessary viper.New() @yegle GH-849
 - authorize: reduce duplicate evaluations in opa policy @travisgroth GH-882
 - envoy: bump envoy to 1.14.2 @desimone GH-894
 - policy: Add consistent route identifier @travisgroth GH-905
 
-## Fixed
+### Fixed
 - xds: use ipv4 address when ipv6 is disabled @calebdoxsey GH-823
 - proxy: only set validation context if trusted_ca is used @calebdoxsey GH-863
 - config: ensure viper ignores `certificates` config field @travisgroth GH-876